requirements.txt
Eel
numpy
Pillow
pip
fpdf
elasticsearch